Refactor: fetchRequest, queries
